applications/house_search/requirements.txt
googlemaps
Flask
PyYAML
chevron
httpx
ijson
imapclient
lxml
mariadb
prefect
prefect-aws
psycopg2-binary
python-dateutil
python-dotenv
pytz
requests
xmltodict
bs4
oauthlib
paramiko
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask_cors